Давай падай

+++ modules/task_1/sokolov_a_sparse_matrix_string_crs/sparse_matrix_string_crs.cpp
@@ -11,8 +11,8 @@ SparseMatrix::SparseMatrix(const Matrix& matrix) {
     rowIndex.reserve(rows + 1);
     rowIndex.push_back(0);
 
-    for (int idx{0}; idx < rows; ++idx) {
-        for (int jdx{0}; jdx < cols; ++jdx) {
+    for (size_t idx{0}; idx < rows; ++idx) {
+        for (size_t jdx{0}; jdx < cols; ++jdx) {
             if (std::fabs(matrix[idx][jdx]) >= tolerance) {
                 value.push_back(matrix[idx][jdx]);
                 colIndex.push_back(jdx);
@@ -23,7 +23,7 @@ SparseMatrix::SparseMatrix(const Matrix& matrix) {
     }
 }
 
-SparseMatrix::SparseMatrix(int _rows, int _cols,
+SparseMatrix::SparseMatrix(size_t _rows, size_t _cols,
                            std::vector<double> _value,
                            std::vector<int>    _colIndex,
                            std::vector<int>    _rowIndex) {
@@ -34,7 +34,7 @@ SparseMatrix::SparseMatrix(int _rows, int _cols,
     rowIndex = _rowIndex;
 }
 
-SparseMatrix::SparseMatrix(int _rows, int _cols, int _elemsCount) {
+SparseMatrix::SparseMatrix(size_t _rows, size_t _cols, int _elemsCount) {
     rows = _rows;
     cols = _cols;
     value.resize(_elemsCount);
@@ -59,7 +59,7 @@ bool SparseMatrix::operator== (const SparseMatrix& matrix) const& {
         std::cout << "Value size are not equal" << std::endl;
         return false;
     }
-    for (int idx{0}; idx < value.size(); ++idx) {
+    for (size_t idx{0}; idx < value.size(); ++idx) {
         if ((std::abs(value[idx] - matrix.value[idx]) > 1e-3)) {
             return false;
             std::cout << "Not Equal! " << value[idx] << " and " << matrix.value[idx]  << std::endl;
@@ -71,13 +71,13 @@ bool SparseMatrix::operator== (const SparseMatrix& matrix) const& {
 Matrix SparseMatrix::SparseToMatrix() {
     Matrix result{};
     result.resize(rows);
-    for (int idx{0}; idx < rows; ++idx) {
+    for (size_t idx{0}; idx < rows; ++idx) {
         result[idx].resize(cols);
     }
     int tmpCols{0};
-    for (int idx{0}; idx < rows; ++idx) {
+    for (size_t idx{0}; idx < rows; ++idx) {
         int tmpRows{rowIndex[idx+1] - rowIndex[idx]};
-        for (int jdx{0}; jdx < cols; jdx++) {
+        for (size_t jdx{0}; jdx < cols; jdx++) {
             if (jdx == colIndex[tmpCols] && tmpRows != 0) {
                 tmpRows--;
                 result[idx][jdx] = value[tmpCols++];
@@ -96,15 +96,15 @@ SparseMatrix SparseMatMul(const SparseMatrix& matrixA, const SparseMatrix& matri
     result.rowIndex.push_back(0);
     std::vector<double> tmpResultRow(matrixA.rows, 0);
 
-    for (int idx{0}; idx < matrixA.rows; ++idx) {
-        for (int jdx{matrixA.rowIndex[idx]}; jdx < matrixA.rowIndex[idx + 1]; ++jdx) {
+    for (size_t idx{0}; idx < matrixA.rows; ++idx) {
+        for (size_t jdx{matrixA.rowIndex[idx]}; jdx < matrixA.rowIndex[idx + 1]; ++jdx) {
             int tmpCol {matrixA.colIndex[jdx]};
 
-            for (int kdx{matrixB.rowIndex[tmpCol]}; kdx < matrixB.rowIndex[tmpCol + 1]; ++kdx) {
+            for (size_t kdx{matrixB.rowIndex[tmpCol]}; kdx < matrixB.rowIndex[tmpCol + 1]; ++kdx) {
                 tmpResultRow[matrixB.colIndex[kdx]] += matrixA.value[jdx] * matrixB.value[kdx];
             }
         }
-        for (int kdx{0}; kdx < matrixA.rows; ++kdx) {
+        for (size_t kdx{0}; kdx < matrixA.rows; ++kdx) {
             if (tmpResultRow[kdx] != 0) {
                 result.value.push_back(tmpResultRow[kdx]);
                 result.colIndex.push_back(kdx);
@@ -118,14 +118,14 @@ SparseMatrix SparseMatMul(const SparseMatrix& matrixA, const SparseMatrix& matri
 
 Matrix MatMul(const Matrix& matrixA, const Matrix& matrixB) {
     Matrix result{matrixA.size()};
-    for (int idx{0}; idx < result.size(); ++idx) {
+    for (size_t idx{0}; idx < result.size(); ++idx) {
         result[idx].resize(matrixB[0].size());
     }
 
-    for (int idx{0}; idx < matrixA.size(); ++idx) {
-        for (int jdx{0}; jdx < matrixB[0].size(); ++jdx) {
+    for (size_t idx{0}; idx < matrixA.size(); ++idx) {
+        for (size_t jdx{0}; jdx < matrixB[0].size(); ++jdx) {
             result[idx][jdx] = 0;
-            for (int kdx{0}; kdx < matrixA[0].size(); ++kdx) {
+            for (size_t kdx{0}; kdx < matrixA[0].size(); ++kdx) {
                 result[idx][jdx] += matrixA[idx][kdx] * matrixB[kdx][jdx];
             }
         }
@@ -134,10 +134,10 @@ Matrix MatMul(const Matrix& matrixA, const Matrix& matrixB) {
 }
 
 // coeff from 0 to 100
-Matrix generateMatrix(const int& rows, const int& cols, const int& coeff) {
+Matrix generateMatrix(const size_t& rows, const size_t& cols, const int& coeff) {
     Matrix result{};
     result.resize(rows);
-    for (int idx{0}; idx < rows; ++idx) {
+    for (size_t idx{0}; idx < rows; ++idx) {
         result[idx].resize(cols);
     }
     std::random_device rd{};
@@ -145,8 +145,8 @@ Matrix generateMatrix(const int& rows, const int& cols, const int& coeff) {
     std::uniform_real_distribution<double> disValue{ 0.0, 10.0 };
     std::uniform_int_distribution<> disProbability {0, 100};
 
-    for (int idx{ 0 }; idx < rows; ++idx) {
-        for (int jdx{ 0 }; jdx < cols; ++jdx) {
+    for (size_t idx{ 0 }; idx < rows; ++idx) {
+        for (size_t jdx{ 0 }; jdx < cols; ++jdx) {
             result[idx][jdx] = 0.0;
             if (disProbability(mt) <= coeff) {
                 result[idx][jdx] = disValue(mt);
@@ -158,15 +158,15 @@ Matrix generateMatrix(const int& rows, const int& cols, const int& coeff) {
 
 void SparseMatrix::printDefault() {
     std::cout << "Value:" << std::endl;
-    for (int idx{ 0 }; idx < value.size(); ++idx) {
+    for (size_t idx{ 0 }; idx < value.size(); ++idx) {
         std::cout << value[idx] << " ";
     }
     std::cout << std::endl << "colIndex:" << std::endl;
-    for (int idx{ 0 }; idx < colIndex.size(); ++idx) {
+    for (size_t idx{ 0 }; idx < colIndex.size(); ++idx) {
         std::cout << colIndex[idx] << " ";
     }
     std::cout << std::endl << "rowIndex:" << std::endl;
-    for (int idx{ 0 }; idx < rowIndex.size(); ++idx) {
+    for (size_t idx{ 0 }; idx < rowIndex.size(); ++idx) {
         std::cout << rowIndex[idx] << " ";
     }
     std::cout << std::endl;
@@ -176,8 +176,8 @@ void SparseMatrix::printMatrix() {
     int tmpValue{ 0 };
     int tmpRows{ 0 };
     int tmpCols{ 0 };
-    for (int idx{ 0 }; idx < rows; ++idx) {
-        for (int jdx{ 0 }; jdx < cols; ++jdx) {
+    for (size_t idx{ 0 }; idx < rows; ++idx) {
+        for (size_t jdx{ 0 }; jdx < cols; ++jdx) {
             if (tmpRows == rowIndex[idx + 1]) {
                 std::cout << "0   ";
             } else if (jdx == colIndex[tmpCols]) {
@@ -194,8 +194,8 @@ void SparseMatrix::printMatrix() {
 }
 
 void print(const Matrix& matrix) {
-    for (int idx{ 0 }; idx < matrix.size(); ++idx) {
-        for (int jdx{ 0 }; jdx < matrix[0].size(); ++jdx) {
+    for (size_t idx{ 0 }; idx < matrix.size(); ++idx) {
+        for (size_t jdx{ 0 }; jdx < matrix[0].size(); ++jdx) {
             std::cout << round(matrix[idx][jdx] * 10) / 10 << " ";
         }
         std::cout << std::endl;
